HH/requirements.txt
asgiref==3.8.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
dbf-light==1.0.0
Django==4.0.5
django-cleanup==6.0.0
django-debug-toolbar==3.4.0
django-filter==21.1
djangorestframework==3.13.1
Faker==0.9.1
Flask==2.1.2
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.3.6
MarkupSafe==2.1.1
mixer==6.1.3
pycbrf==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2024.1
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.5.1
text-unidecode==1.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.9
Werkzeug==2.1.2
